build: remove no longer used scripts
This commit removes scripts which have been replaced with Bazel.
