+++ package.json
@@ -19,9 +19,8 @@
   ],
   "scripts": {
     "admin": "node ./bin/devkit-admin",
-    "bazel:test": "bazel test //packages/...",
+    "test": "bazel test //packages/...",
     "build": "node ./bin/devkit-admin build",
-    "build:bazel": "node ./bin/devkit-admin build-bazel",
     "build-tsc": "tsc -p tsconfig.json",
     "lint": "eslint --cache --max-warnings=0 \"**/*.ts\"",
     "ng-dev": "cross-env TS_NODE_PROJECT=$PWD/.ng-dev/tsconfig.json TS_NODE_TRANSPILE_ONLY=1 node --no-warnings --loader ts-node/esm node_modules/@angular/ng-dev/bundles/cli.mjs",
