+++ scripts/json-help.ts
@@ -38,7 +38,7 @@ export default async function ({ temporaryProjectRoot }: JsonHelpOptions, logger
 
   await fs.mkdir(helpOutputRoot);
 
-  const runNgCommandJsonHelp = (args: string[]): Promise<JsonHelp> => {
+  const runNgCommandJsonHelp = async (args: string[]): Promise<JsonHelp> => {
     const { stdout, status } = spawnSync(ngPath, [...args, '--json-help', '--help'], {
       cwd: newProjectRoot,
       maxBuffer: 200_0000,
@@ -46,10 +46,14 @@ export default async function ({ temporaryProjectRoot }: JsonHelpOptions, logger
     });
 
     if (status === 0) {
-      return Promise.resolve(JSON.parse(stdout.toString().trim()));
-    } else {
-      throw new Error(`Command failed: ${ngPath} ${args.map((x) => JSON.stringify(x)).join(', ')}`);
+      try {
+        return JSON.parse(stdout.toString().trim());
+      } catch (e) {
+        logger.error(`${e}`);
+      }
     }
+
+    throw new Error(`Command failed: ${ngPath} ${args.map((x) => JSON.stringify(x)).join(', ')}.`);
   };
 
   const { subcommands: commands = [] } = await runNgCommandJsonHelp([]);
